refactor(fetch_azure): split modules and fix runtime cancellation
Address PR review feedback:
- move AzureHttpClient into a client module
- move the runtime into a runtime module and rename SpawnerRuntime to Runtime
- fix cancellation: spawn tasks wrapped in futures::future::Abortable and
  abort via AbortHandle so aborting wakes pending waiters, instead of the
  flag-based approach that left parked awaiters hanging

lib.rs is now a thin module root; client.rs and runtime.rs are each at 100%
coverage.
