+++ crates/fetch_azure/CHANGELOG.md
@@ -8,5 +8,5 @@
     Oxidizer stack:
     - `FetchHttpClient` implements `azure_core::http::HttpClient` on top of a
       `fetch::HttpClient` transport.
-    - `SpawnerRuntime` implements `azure_core::async_runtime::AsyncRuntime` on top
+    - `Runtime` implements `azure_core::async_runtime::AsyncRuntime` on top
       of an `anyspawn::Spawner` (spawning) and a `tick::Clock` (sleeping).
